refactor(shared): move sendOrderEmail to server api
